% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{histo4param_values}
\alias{histo4param_values}
\title{histo4param_values}
\usage{
histo4param_values(distr, method, param)
}
\arguments{
\item{distr}{}

\item{method}{}

\item{param}{}
}
\description{
Histogram to go with the parameter mapping
}

