% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{norway_map4param_values}
\alias{norway_map4param_values}
\title{norway_map4param_values}
\usage{
norway_map4param_values(distr, method, param)
}
\arguments{
\item{distr}{}

\item{method}{}

\item{param}{}
}
\description{
Function to map the fitted parameter values as color scale for each station
}

