% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{norway_map4server}
\alias{norway_map4server}
\title{norway_map4server}
\usage{
norway_map4server(selected.station)
}
\arguments{
\item{selected.station}{}
}
\description{
MAP OF NORWAY WITH COLOR CHANGING ACCORDING TO THE NUMBER OF DATA.
}

