% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_plotting.R
\name{plot4server}
\alias{plot4server}
\title{plot4server}
\usage{
plot4server(dat, param, distr.index = 1)
}
\arguments{
\item{dat}{}

\item{param}{}

\item{distr.index}{}
}
\value{
Returns nothing else than a plot, saves nothing
}
\description{
plot fitted probability density function to estimated empirical pdf
}

