% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_plotting.R
\name{plot4server_cdf}
\alias{plot4server_cdf}
\title{plot4server_cdf}
\usage{
plot4server_cdf(dat, param, distr = 1)
}
\arguments{
\item{dat}{}

\item{param}{}

\item{distr}{}
}
\value{
Returns nothing else than a plot, saves nothing
}
\description{
Plot estimated and empirical cumulative distribution function
}

