% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rperiods_plotting.R
\name{plot4server_gof.rlevels}
\alias{plot4server_gof.rlevels}
\title{plot4server_gof.rlevels}
\usage{
plot4server_gof.rlevels(station, gof, r.period, quantile)
}
\arguments{
\item{station}{}

\item{gof}{}

\item{r.period}{}

\item{quantile}{}
}
\description{
plot4server_gof.rlevels
}

