% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_plotting.R
\name{plot4server_gof_averaged}
\alias{plot4server_gof_averaged}
\title{plot4server_gof_averaged}
\usage{
plot4server_gof_averaged(gof, min_years, max_years,
  stations2ave = stations2average(min_years, max_years))
}
\arguments{
\item{gof}{}

\item{min_years}{}

\item{max_years}{}
}
\description{
function like plot4server_gof but averaging over all stations filling the min/max years criteria
}

