% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rperiods_plotting.R
\name{plot4server_nt}
\alias{plot4server_nt}
\title{plot4server_nt}
\usage{
plot4server_nt(station, gof, r.period)
}
\arguments{
\item{station}{}

\item{gof}{}

\item{r.period}{}
}
\description{
NT is plotted differently than the other r.levels indices 
because it doens't have the extra "few_quantiles" dimension
}

