% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_plotting.R
\name{plot4server_qq}
\alias{plot4server_qq}
\title{plot4server_qq}
\usage{
plot4server_qq(dat, param, distr = 1)
}
\arguments{
\item{dat}{}

\item{param}{}

\item{distr}{}
}
\value{
Returns nothing else than a plot else than a plot, saves nothing
}
\description{
QQ plot of empiricial against modelled
}

