% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_plotting.R
\name{plot4server_rlevel}
\alias{plot4server_rlevel}
\title{plot4server_rlevel}
\usage{
plot4server_rlevel(dat, param, distr.index = 1)
}
\arguments{
\item{dat}{}

\item{param}{}

\item{distr.index}{}
}
\value{
Returns nothing else than a plot, saves nothing
}
\description{
Plots return levels
}

