% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rperiods_plotting.R
\name{plot4server_rlevels_coeffvar}
\alias{plot4server_rlevels_coeffvar}
\title{plot4server_rlevels_coeffvar}
\usage{
plot4server_rlevels_coeffvar(station, gof, r.period)
}
\arguments{
\item{station}{}

\item{gof}{}

\item{r.period}{}
}
\description{
Plotting the coefficient of variation of return levels
}

