% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rperiods_plotting.R
\name{plot4server_rlevels_coeffvar_ave}
\alias{plot4server_rlevels_coeffvar_ave}
\title{plot4server_rlevels_coeffvar_ave}
\usage{
plot4server_rlevels_coeffvar_ave(gof, r.period, min_years, max_years,
  stations2ave = stations2average(min_years, max_years))
}
\arguments{
\item{gof}{}

\item{r.period}{}

\item{min_years}{}

\item{max_years}{}
}
\description{
plot4server_rlevels_coeffvar_ave
}

