% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata_plotting.R
\name{qdata_boxplot_OLD}
\alias{qdata_boxplot_OLD}
\title{qdata_boxplot_OLD}
\usage{
qdata_boxplot_OLD(min_years, max_years, min_height, max_height)
}
\arguments{
\item{min_years}{}

\item{max_years}{}

\item{min_height}{}

\item{max_height}{}
}
\description{
qdata_boxplot_OLD
}

